Move spend_cache cached operations to its own module
